Extracted context-aware utilities - moved context-aware helper methods to their own static utility class in a new script - organized imports
