+++ Runtime/Scripts/Behaviors/P300ControllerBehavior.cs
@@ -1,16 +1,18 @@
-using BCIEssentials.StimulusObjects;
 using System.Collections;
-using UnityEngine;
+using System.Collections.Generic;
 using System;
 using System.Linq;
+using UnityEngine;
+using Random = System.Random;
 using BCIEssentials.Controllers;
 using BCIEssentials.Utilities;
-using Random = System.Random;
-using System.Collections.Generic;
-using UnityEditor.EditorTools;
+using BCIEssentials.StimulusObjects;
+
 
 namespace BCIEssentials.ControllerBehaviors
 {
+    using static ContextAwareUtilities;
+
     public class P300ControllerBehavior : BCIControllerBehavior
     {
         public override BCIBehaviorType BehaviorType => BCIBehaviorType.P300;
@@ -132,7 +134,7 @@ namespace BCIEssentials.ControllerBehaviors
                 //Refresh the List of available SPO Objects. This will update the _validGOs list.
                 PopulateObjectList();
                 //Now get the Graph set for the TSP
-                int[] stimOrder = CalculateGraphTSP(_validGOs);
+                int[] stimOrder = CalculateGraphTSP(_validGOs, ref lastTourEndNode);
 
                 for (int i = 0; i < stimOrder.Length; i++)
                 {
@@ -497,138 +499,6 @@ namespace BCIEssentials.ControllerBehaviors
             return visibleGOs;
         }
 
-        public int[] CalculateGraphTSP(List<GameObject> nodes, bool debugPrint = false)
-        {
-            //Get the world points of each item with respect to the camera
-            // var cameraTranfsorm = Camera.main.transform;
-            List<Vector2> correctedNodePositions = CalculateOffsetFromCamera(nodes, Camera.main);
-            int numNodes = nodes.Count;
-            float[,] objectWeights = new float[numNodes, numNodes];
-        
-            foreach (var node in nodes)
-            {
-                //use Vector3.Angle to get the angle between every object in the scene. Store this as weights in a graph
-                //This is a symmetric matrix, so we only need to calculate the upper triangle
-                for (int i = 0; i < numNodes; i++)
-                {
-                    if (i == nodes.IndexOf(node))
-                    {
-                        objectWeights[nodes.IndexOf(node), i] = 0;
-                    }
-                    else
-                    {
-                        objectWeights[nodes.IndexOf(node), i] = Vector3.Angle(correctedNodePositions[nodes.IndexOf(node)], correctedNodePositions[i]);
-                    }
-                }
-            }
-
-            if (debugPrint)
-            {
-                //Print the weights in the upper triangle matrix
-                for (int i = 0; i < numNodes; i++)
-                {
-                    for (int j = 0; j < numNodes; j++)
-                    {
-                        if (j >= i && objectWeights[i, j] != 0)
-                        {
-                            Debug.Log($"Angle (weight) between {nodes[i].name} and {nodes[j].name}: {objectWeights[i, j]}");
-                        }
-                    }
-                }
-            }
-
-
-            GraphUtilities tsp = new GraphUtilities();
-            
-            var startNode = UnityEngine.Random.Range(0, numNodes);
-            //Make sure the start node and last node of the tour are not the same
-            if(startNode == lastTourEndNode)
-            {
-                //chose a different start node
-                startNode = (startNode+1) % numNodes;
-            }
-          
-            var tour = tsp.SolveModifiedTSP(objectWeights, startNode);
-            lastTourEndNode = tour[tour.Count - 1];
-
-            if(debugPrint)
-            {
-                Debug.Log("The start node is " + startNode.ToString());
-                            //Print out the tour
-                for (int i = 0; i < tour.Count; i++)
-                {
-                    Debug.Log("The tour is " + tour[i].ToString());
-                }
-                Debug.Log("Tour length is: " + tsp.CalculateTourLength(tour));
-            }
-
-            int[] tourArray = tour.ToArray();
-            return tourArray;
-  
-
-            //Use the graph utilities to do the modified TSP.      
-
-        }
-
-        public (int[] subset1, int[] subset2) CalculateGraphPartition(List<GameObject> nodes)
-        {
-            // Get 2D screen positions
-            List<Vector2> screenPositions = CalculateOffsetFromCamera(nodes, Camera.main);
-            int numNodes = nodes.Count;
-            float[,] objectWeights = new float[numNodes, numNodes];
-
-            foreach (var node in nodes)
-            {
-                int i = nodes.IndexOf(node);
-                for (int j = 0; j < numNodes; j++)
-                {
-                    if (i == j)
-                    {
-                        objectWeights[i, j] = 0;
-                    }
-                    else
-                    {
-                        // Calculate 2D screen-space distance
-                        float distance = Vector2.Distance(
-                            screenPositions[i], 
-                            screenPositions[j]
-                        );
-                        // Convert distance to weight (inverse relationship)
-                        objectWeights[i, j] = 1.0f / (distance + 1.0f);
-                    }
-                }
-            }
-
-            var lpPart = new GraphUtilities();
-            return lpPart.LaplaceGP(objectWeights);
-        }
-
-        public List<Vector2> CalculateOffsetFromCamera(List<GameObject> goList, Camera myCamera)
-        {
-            List<Vector2> screenPositions = new();
-            foreach (var obj in goList)
-            {
-                if (obj.layer == 5 || obj.TryGetComponent<RectTransform>(out var rectT))
-                {
-                    // UI Elements - get screen position directly from RectTransform
-                    RectTransform tempRT = obj.GetComponent<RectTransform>();
-                    RectTransformUtility.ScreenPointToWorldPointInRectangle(
-                        tempRT,
-                        tempRT.position,
-                        myCamera,
-                        out Vector3 screenPos
-                    );
-                    screenPositions.Add(new Vector2(screenPos.x, screenPos.y));
-                }
-                else
-                {
-                    // 3D Objects - project to screen space
-                    Vector3 screenPos = myCamera.WorldToScreenPoint(obj.transform.position);
-                    screenPositions.Add(screenPos);
-                }
-            }
-            return screenPositions;
-        }
 
         protected override void SendTrialStartedMarker()
         {
